+++ app/src/main/java/com/example/taskyapplication/auth/presentation/AccountCreationScreen.kt
@@ -3,40 +3,24 @@ package com.example.taskyapplication.auth.presentation
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.Check
 import androidx.compose.material3.Icon
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.autofill.ContentType
-import androidx.compose.ui.focus.onFocusChanged
 import androidx.compose.ui.platform.LocalAutofillManager
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.semantics.contentType
-import androidx.compose.ui.semantics.semantics
-import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.example.taskyapplication.R
 import com.example.taskyapplication.TaskyBaseScreen
 import com.example.taskyapplication.auth.domain.RegisterUserState
 import com.example.taskyapplication.auth.presentation.components.AuthScreenFooter
 import com.example.taskyapplication.auth.presentation.components.AuthorizationCtaButton
 import com.example.taskyapplication.auth.presentation.components.BaseInputField
 import com.example.taskyapplication.auth.presentation.components.PasswordTextField
-import com.example.taskyapplication.auth.presentation.components.UserInfoTextField
 import com.example.taskyapplication.auth.presentation.utils.AuthScreenTitle
 import com.example.taskyapplication.ui.theme.TaskyDesignSystem.Companion.taskyColors
-import com.example.taskyapplication.ui.theme.TaskyTypography
 
 @Composable
 fun RegisterRoot(
