+++ app/src/main/java/com/example/taskyapplication/auth/presentation/components/BaseTextField.kt
@@ -10,12 +10,15 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.BasicText
 import androidx.compose.foundation.text.BasicTextField
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.text.input.TextFieldLineLimits
 import androidx.compose.foundation.text.input.TextFieldState
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.Check
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -28,8 +31,12 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.focus.onFocusChanged
 import androidx.compose.ui.graphics.SolidColor
+import androidx.compose.ui.text.LinkAnnotation
+import androidx.compose.ui.text.TextLinkStyles
+import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.withLink
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.example.taskyapplication.ui.theme.TaskyDesignSystem.Companion.taskyColors
@@ -114,6 +121,76 @@ fun BaseInputField(
     }
 }
 
+@Composable
+fun AuthorizationCtaButton(
+    modifier: Modifier = Modifier,
+    buttonText: String,
+    isButtonEnabled: Boolean = false,
+    onButtonClick: () -> Unit = { },
+) {
+    Button(
+        modifier = modifier
+            .fillMaxWidth()
+            .height(96.dp)
+            .padding(start = 20.dp, end = 20.dp, top = 16.dp, bottom = 16.dp)
+            .clip(RoundedCornerShape(25.dp)),
+        onClick = onButtonClick,
+        enabled = isButtonEnabled,
+        colors = ButtonDefaults.buttonColors(
+            containerColor = taskyColors.primary,
+            contentColor = taskyColors.onPrimary,
+            disabledContainerColor = taskyColors.primary,
+        )
+    ) {
+        Text(
+            text = buttonText.uppercase(),
+            color = taskyColors.onPrimary,
+            style = TaskyTypography.labelMedium
+        )
+    }
+}
+
+@Composable
+fun AuthScreenFooter(
+    modifier: Modifier = Modifier,
+    navigateToScreen: () -> Unit,
+    accountRegisteredPrompt: String,
+    loginOrSignupPrompt: String
+) {
+    Spacer(
+        modifier = modifier
+            .fillMaxWidth()
+            .height(16.dp)
+    )
+    Row(
+        modifier = Modifier
+            .fillMaxWidth(),
+        horizontalArrangement = Arrangement.Center
+    ) {
+        BasicText(
+            text = buildAnnotatedString {
+                append("${accountRegisteredPrompt}  ".uppercase())
+                withLink(
+                    LinkAnnotation.Clickable(
+                        tag = "",
+                        styles = TextLinkStyles(
+                            style = TaskyTypography.labelSmall.toSpanStyle().copy(
+                                color = taskyColors.link
+                            )
+                        ),
+                        linkInteractionListener = { navigateToScreen() }
+                    ),
+                ) {
+                    append(loginOrSignupPrompt)
+                }
+            },
+            style = TaskyTypography.labelSmall.copy(
+                color = taskyColors.onSurfaceVariant,
+            )
+        )
+    }
+}
+
 @Preview(showBackground = true, backgroundColor = 0xFF00FF00)
 @Composable
 fun PlaygroundPreview() {
@@ -131,5 +208,13 @@ fun PlaygroundPreview() {
                     .padding(8.dp)
             )
         }
+    )
+}
+
+@Preview(showBackground = true, backgroundColor = 0xFF00FF00)
+@Composable
+fun CtaButtonPreview() {
+    AuthorizationCtaButton(
+        buttonText = "Sign In"
     )
 }
